backend/requirements.txt
uvicorn
fastapi
selenium
selenium-stealth
pandas
webdriver-manager
python-multipart
python-decouple

